requirements.txt
tortoise-orm
aerich
vk_api
asyncpg
aiohttp~=3.7.4.post0